requirement.txt

discord.py==2.3.2
python-dotenv==1.0.1
aiosqlite==0.19.0   # async sqlite for XP system
